Move IdentityTypeDescriptor to impl package.
