+++ common/src/main/kotlin/com/github/fkrull/konflin/typedescriptors/impl/IdentityTypeDescriptor.kt
@@ -1,5 +1,7 @@
-package com.github.fkrull.konflin.typedescriptors
+package com.github.fkrull.konflin.typedescriptors.impl
 
+import com.github.fkrull.konflin.typedescriptors.ConfigType
+import com.github.fkrull.konflin.typedescriptors.TypeDescriptor
 import kotlin.reflect.KClass
 
 internal class IdentityTypeDescriptor<Type : Any>(override val configType: ConfigType<Type>) : TypeDescriptor<Type, Type> {
